pyproject.toml
[project]
name = "obs-cam-focus-switcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "obsws-python>=1.6.0",
    "pywin32>=306",
    "psutil>=5.9.0",
    "simpleobsws>=1.4.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "pylint>=2.17.0",
]
